Extract component definition for tested components
